Comments working, simple httprequest, better GC
comments now working again. ```innerHTML= null ``` was effecting it. oops.

Moved some ```XMLHTTPrequests``` to ```httprequest()``` to allow for smaller filesize and ease of maintaining.

now everything replaced or inserted is garbage collected when viewer is out of the window.

also now a 30px margin at the bottom to give a floating effect because i want it that way.